Added and implemented useFilter hook
